\name{gaussian_kernel}
\alias{gaussian_kernel}

\title{
Helper function for kernel smoother
}
\description{
Gaussian kernel used for kernel smoother.
}
\usage{
gaussian_kernel(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
x
}
}

\value{
Gaussian kernel applied to x
}

\author{
Emily Hsiao
}
\keyword{internal}
