\name{S}
\alias{S}

\title{
Helper function
}
\description{
Helper function for monotonicity test.
}
\usage{
S(a, b, r, s, X, Y)
}

\arguments{
  \item{a}{
Value of a (regression coefficient)
}
  \item{b}{
Value of b (regression coefficient)
}
  \item{r}{
Index to start summation
}
  \item{s}{
Index to end summation
}
  \item{X}{
Vector of X values
}
  \item{Y}{
Vector of Y values
}
}

\value{
Mean-squared error
}

\author{
Emily Hsiao
}

\keyword{internal}
